requirements.txt
# QPP-Fusion-RAG Dependencies
# Research-grade IR/RAG evaluation pipeline

# =============================================================================
# Core IR/Retrieval
# =============================================================================
pyterrier>=0.10.0
pyterrier-dr>=0.4.0
pyserini>=0.22.0
faiss-cpu>=1.7.4

# =============================================================================
# Evaluation (Research-Grade)
# =============================================================================
ir_measures>=0.3.3
ranx>=0.3.16
evaluate>=0.4.0
rouge_score>=0.1.2
nltk>=3.8.0
absl-py>=2.0.0

# =============================================================================
# ML Models
# =============================================================================
lightgbm>=4.0.0
torch>=2.0.0
sentence-transformers>=2.2.0

# =============================================================================
# Data Processing
# =============================================================================
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# =============================================================================
# HuggingFace
# =============================================================================
transformers>=4.35.0
datasets>=2.14.0
huggingface-hub>=0.19.0

# =============================================================================
# Utilities
# =============================================================================
requests>=2.31.0
tqdm>=4.66.0

# =============================================================================
# Optional: Java QPP Bridge
# =============================================================================
# Requires: Java 11+, gson-2.11.0.jar in lib/

